GolgeBahcesi: address PR review feedback
- Use HttpUrl.pathSegments instead of string splitting
- Add versionId = 2 for URL structure change
- Move filter classes to Filters.kt
